@@ -18,8 +18,9 @@
     },
     "bin": ["bin/phpfci"],
     "require": {
-        "php": ">=7.2",
+        "php": ">=7.4",
         "ext-dom": "*",
+        "ext-json": "*",
         "ext-libxml": "*",
         "ext-xmlwriter": "*",
         "symfony/console": "^5.1"
@@ -28,7 +29,7 @@
         "roave/security-advisories": "dev-master",
         "squizlabs/php_codesniffer": "^3.5",
         "phpmd/phpmd": "@stable",
-        "phpunit/phpunit": "8.5.* || 9.4.*",
+        "phpunit/phpunit": "9.5.*",
         "phpstan/phpstan-phpunit": "0.12.*",
         "phpstan/phpstan-strict-rules": "0.12.*",
         "phpstan/extension-installer": "1.0.*",
